requirements.txt
Flask
Flask-SQLAlchemy
APIFlask
flask-jwt-extended
cwb-ccc
flask-login
pytest
